dom/tests/mochitest/pointerlock/test_pointerlock-api.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
  <head>
    <title>Test for Bug 633602</title>
    <script type="application/javascript"
      src="/tests/SimpleTest/SimpleTest.js"></script>
    <script type="application/javascript"
      src="/tests/SimpleTest/EventUtils.js"></script>
    <link rel="stylesheet" type="text/css"
          href="/tests/SimpleTest/test.css"/>
  </head>
  <body>
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
      Mozilla Bug 633602
    </a>
    <div id="content">
    </div>
    <pre id="test">
      <script type="application/javascript">
        SimpleTest.waitForExplicitFinish();
        // Ensure the full-screen api is enabled,
        // and will be disabled on test exit.
        SpecialPowers.setBoolPref("full-screen-api.enabled", true);

        // Disable the requirement for trusted contexts only,
        // so the tests are easier to write.
        SpecialPowers.setBoolPref("full-screen-api.allow-trusted-requests-only",
                                  false);

        // Run the tests which go full-screen in new window, as Mochitests
        // normally run in an iframe, which by default will not have the
        // mozallowfullscreen attribute set, so full-screen won't work.
        var gTestFiles = [
          "file_childIframe.html",
          "file_doubleLock.html",
          "file_escapeKey.html",
          "file_infiniteMovement.html",
          "file_locksvgelement.html",
          "file_loseFocusWindow.html",
          "file_movementXY.html",
          "file_nestedFullScreen.html",
          "file_pointerlock-api.html",
          "file_pointerlockerror.html",
          "file_pointerLockPref.html",
          "file_removedFromDOM.html",
          "file_retargetMouseEvents.html",
          "file_screenClientXYConst.html",
          "file_suppressSomeMouseEvents.html",
          "file_targetOutOfFocus.html",
          "file_withoutDOM.html"
        ];

        var gTestWindow = null;
        var gTestIndex = 0;

        // TODO: if ever we remove these checks for XP and Lion, we should do the same
        // in content/html/content/test/test_fullscreen-api.html, which uses the same pattern.
        const isWinXP = navigator.userAgent.indexOf("Windows NT 5.1") != -1;
        const isOSXLion = navigator.userAgent.indexOf("Mac OS X 10.7") != -1;

        function nextTest() {
          if (isWinXP) {
            todo(false, "Can't reliably run full-screen tests on Windows XP due to bug 704010");
            SimpleTest.finish();
            return;
          }
          if (gTestWindow) {
            gTestWindow.close();
            if (isOSXLion) {
              // On OS X Lion, tests cause problems. Timeouts are a bad way to get around
              // the problem and may lead to future [orange], but they are the only option
              // at this point.
              SimpleTest.waitForFocus(function() { setTimeout(runNextTest, 3000); });
              return;
            }
          }
          runNextTest();
        }

        function runNextTest() {
          if (gTestIndex < gTestFiles.length) {
            gTestWindow = window.open(gTestFiles[gTestIndex], "", "width=500,height=500");
            // We'll wait for the window to load, then make sure our window is refocused
            // before starting the test, which will get kicked off on "focus".
            // This ensures that we're essentially back on the primary "desktop" on
            // OS X Lion before we run the test.
            gTestWindow.addEventListener("load", function onload() {
              gTestWindow.removeEventListener("load", onload, false);
              SimpleTest.waitForFocus(function() {
                SimpleTest.waitForFocus(gTestWindow.start, gTestWindow);
              });
            }, false);
            gTestIndex++;
          } else {
            SpecialPowers.clearUserPref("full-screen-api.enabled");
            SpecialPowers.clearUserPref("full-screen-api.allow-trusted-requests-only");
            SimpleTest.finish();
          }
        }

        addLoadEvent(nextTest);
      </script>
    </pre>
  </body>
</html>
